// src/app/models/user.ts
export class User {
  id: string;
  _id: string;
  firstName: string;
  lastName: string;
  email: string;
  password: string;
  dob: string;
  gender: string;
  photoId: string;
  token: string;
  isAdmin: boolean;
  isActive: boolean;
  photoUrl: any;
  friendStatus: string;
  city: string;
  state: string;
  country: string;
  phone: string;
}